Use type definition for bitboard

+++ internal/tables/knight_tables_internal_test.go
@@ -8,7 +8,7 @@ import (
 )
 
 func TestMaskKnightAttacks(t *testing.T) {
-	testCases := map[int]int{
+	testCases := map[int]uint64{
 		sq.E4: 11333767002587136,   // central
 		sq.G4: 45053588738670592,   // g-file
 		sq.B5: 5531918402816,       // b-file
@@ -23,7 +23,7 @@ func TestMaskKnightAttacks(t *testing.T) {
 
 	for square, expected := range testCases {
 		actual := maskKnightAttacks(square)
-		if actual != uint64(expected) {
+		if uint64(actual) != expected {
 			bitboard.PrintBoard(actual)
 			t.Errorf("Getting knight attacks for %s, expected %d, got %d", sq.Stringify(square), expected, actual)
 		}
